Create directory structure to organize variations
