+++ model.py
@@ -15,304 +15,11 @@ import torch
 import torch.nn as nn
 from torch.nn import functional as F
 
-class LayerNorm(nn.Module):
-    """ LayerNorm but with an optional bias. PyTorch doesn't support simply bias=False """
-
-    def __init__(self, ndim, bias):
-        super().__init__()
-        self.weight = nn.Parameter(torch.ones(ndim))
-        self.bias = nn.Parameter(torch.zeros(ndim)) if bias else None
-
-    def forward(self, input):
-        return F.layer_norm(input, self.weight.shape, self.weight, self.bias, 1e-5)
-
-class RMSNorm(nn.Module):
-    """RMS Normalization"""
-
-    def __init__(self, ndim):
-        super().__init__()
-        self.gain = nn.Parameter(torch.ones(ndim))
-
-    def forward(self, x):
-        rms = x.norm(2, dim=-1, keepdim=True) / math.sqrt(x.size(-1))
-        return x / rms * self.gain
-
-# Softmax base 2, with option to remove max subtraction
-class Softermax(nn.Module):
-    """ Base-2 Softmax with option to remove max subtraction"""
-    def __init__(self, config, dim=-1):
-        super().__init__()
-        self.dim = dim
-        self.subtract_max = config.softermax_use_xmax
-
-    def forward(self, x):
-        if self.subtract_max:
-            max_x = x.max(dim=self.dim, keepdim=True).values
-            x = x - max_x
-        e_x = torch.pow(2.0, x)
-        return e_x / e_x.sum(dim=self.dim, keepdim=True)
-
-# Softmax variation with learnable constant parameters for xmax and denominator
-class Constantmax(nn.Module):
-    """ Constant learnable parameters for xmax and denominator """
-    def __init__(self, config, dim=-1):
-        super().__init__()
-
-        # learnable 'xmax' - beta
-        self.beta = nn.Parameter(torch.Tensor([config.constantmax_initial_beta]))
-
-        # denominator - gamma
-        self.gamma = nn.Parameter(torch.Tensor([config.constantmax_initial_gamma]))
-
-        # Set the base of the exponent
-        if config.constantmax_use_euler_base:
-          # equivalent to 'e'
-          self.constantmax_base = torch.exp(torch.tensor(1.0))
-        else:
-          self.constantmax_base = config.constantmax_base
-
-    def forward(self, x):
-        x = x - self.beta
-        e_x = torch.pow(self.constantmax_base, x)
-        return e_x / self.gamma
-
-# Constantmax Quantized
-
-## Quantization Methods Utilized for Separate Forward and Backward Passes
-def quantize(tensor,scale):
-    tensor = tensor.mul(scale)
-    tensor = torch.round(tensor)
-    return tensor
-def dequantize(tensor,scale):
-    tensor = tensor.div(scale)
-    return tensor
-
-## helper class for Constantmax_quan
-class const_quan(torch.autograd.Function):
-    """Simulates error caused by quantization. Uses Straight-Through Estimator for Back prop"""
-    @staticmethod
-    def forward(ctx, beta=None, gamma=None):
-        #scaling factor for beta and gamma while doing quantization
-        scale_beta=100 #scaling factor for quantization, should make it as parameter
-        scale_gamma=10
-        beta = quantize(beta, scale_beta)
-        gamma = quantize(gamma, scale_gamma)
-        return dequantize(beta, scale_beta),dequantize(gamma,scale_gamma)
-
-    @staticmethod
-    def backward(ctx, grad_gamma, grad_beta):
-        return grad_gamma, grad_beta
-
-_const_quan=const_quan.apply
-
-class Constantmax_quan(nn.Module):
-    """ Base-e Softmax with option to remove max subtraction"""
-    def __init__(self, config, dim=-1):
-        super().__init__()
-        self.dim = dim
-
-        # demonimator - gamma
-        self.gamma = nn.Parameter(torch.Tensor([config.constantmax_initial_gamma]))
-
-        # learnable 'xmax' - beta
-        self.beta = nn.Parameter(torch.Tensor([config.constantmax_initial_beta]))
-
-        self.fake_beta = None
-        self.fake_gamma = None
-
-    def forward(self, x):
-        if self.training:
-            #print('fake_beta:', self.fake_beta)
-            #print('fake_gamma:', self.fake_gamma)
-            self.fake_beta, self.fake_gamma=_const_quan(self.beta, self.gamma)
-            x = x - self.fake_beta
-            e_x = torch.exp(x)
-            return e_x / self.fake_gamma
-        else:
-            scale_beta=100 #scaling factor for quantization, should make it as parameter
-            scale_gamma=10
-            x = x - dequantize(quantize(self.beta,scale_beta), scale_beta)
-            e_x = torch.exp(x)
-            return e_x/dequantize(quantize(self.gamma,scale_gamma), scale_gamma)
-
-# Like softermax, but parameterized to permit exploration of bases greater than 2
-class Strongermax(nn.Module):
-    """ Softmax with ability to increase to 'stronger' bases """
-    def __init__(self, config, dim=-1):
-        super().__init__()
-        self.dim = dim
-        self.strength = config.strongermax_strength
-        self.subtract_max = config.softermax_use_xmax
-
-    def forward(self, x):
-        if self.subtract_max:
-            max_x = x.max(dim=self.dim, keepdim=True).values
-            x = x - max_x
-        e_x = torch.pow(self.strength, x)
-        return e_x / e_x.sum(dim=self.dim, keepdim=True)
-
-# Using polynomial instead of exponential for Softmax separation non-linearity
-class Polymax(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-
-        assert(config.polymax_x_intercept < 0) # ensure x_intercept is strictly left of the y-axis
-
-        self.x_intercept = config.polymax_x_intercept # where to transition from y=0 to m*x+b
-        self.y_intercept = config.polymax_y_intercept # where teh graph crosses y-axis
-
-        self.power = config.polymax_power
-        self.divisor = config.polymax_divisor
-
-    def forward(self, x):
-        # Overview:
-        # Flat section:       -inf < x < x_intercept
-        # Linear section:     x_intercept <= x <= 0
-        # Polynomial section: 0 < x < inf
-
-        # Flat section
-        flat_piece = torch.where(x < self.x_intercept, torch.tensor(0.0, device=x.device), torch.tensor(0.0, device=x.device))
-
-        # Linear section
-        m = self.y_intercept/self.x_intercept # aka 'slope', also x intercept !=0
-        b = self.y_intercept
-        linear_piece = torch.where((x >= self.x_intercept) & (x <= 0), m * x + b, torch.tensor(0.0, device=x.device))
-
-        # Polynomial section
-        poly_piece = torch.where(x > 0, x**self.power + self.y_intercept, torch.tensor(0.0, device=x.device))
-
-        # Combine sections
-        return (poly_piece + linear_piece + flat_piece)/self.divisor
-
-# SigSoftmax from https://arxiv.org/abs/1805.10829
-class SigSoftmax(nn.Module):
-    def __init__(self, config, dim=-1):
-        super().__init__()
-        self.dim = dim
-
-    def forward(self, inputs):
-        exp_x = torch.exp(inputs)
-        sig_x = torch.sigmoid(inputs)
-
-        numerator = exp_x * sig_x
-        denominator = torch.sum(exp_x * sig_x, dim=self.dim, keepdim=True)
-
-        return numerator / denominator
-
-# SigSoftmax from https://arxiv.org/abs/1805.10829
-class SigSoftmax(nn.Module):
-    """ Softmax variant based on arxiv 1805.10829 with added handles for base """
-    def __init__(self, config, dim=-1):
-        super().__init__()
-        self.dim = dim
-        self.base = config.sigsoftmax_base
-
-        # Set the base of the exponent
-        if config.sigsoftmax_use_euler_base:
-          # equivalent to 'e'
-          self.sigsoftmax_base = torch.exp(torch.tensor(1.0))
-        else:
-          # custom base
-          self.sigsoftmaxmax_base = config.sigsoftmax_base
-
-    def forward(self, inputs):
-
-        # Set exponent
-        exp_x = torch.pow(self.sigsoftmax_base, inputs)
-
-        # Similarly set sigmoid approximation
-        sig_x = 1 / (1 + torch.pow(self.sigsoftmax_base, -inputs))
-
-        # calculation of numerator and denominator
-        numerator = exp_x * sig_x
-        denominator = torch.sum(exp_x * sig_x, dim=self.dim, keepdim=True)
-
-        return numerator / denominator
-
-# POSITIONAL EMBEDDINGS
-
-class RotaryEmbedding(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        self.dim = config.n_embd
-
-        # Register frequencies directly as buffers
-        self.register_buffer('freq_left', (10000 ** (torch.arange(0, self.dim//2).float() / self.dim//2)))
-        self.register_buffer('freq_right',(10000 ** (torch.arange(0, self.dim//2).float() / self.dim//2)))
-
-    def forward(self, x):
-        seq_len = x.shape[-2]
-        device = x.device
-
-        t = torch.arange(seq_len, device=device)
-
-        # Get separate frequencies for left and right
-        freqs_left = torch.einsum('i,j->ij', t, self.freq_left)
-        freqs_right = torch.einsum('i,j->ij', t, self.freq_right)
-
-        # Apply frequencies
-        x_left, x_right = x[..., :self.dim//2], x[..., self.dim//2:]
-        x_left = x_left * freqs_left.cos() - x_right * freqs_left.sin()
-        x_right = x_left * freqs_right.sin() + x_right * freqs_right.cos()
-
-        # Combine the left and right parts back
-        x = torch.cat([x_left, x_right], dim=-1)
-
-        return x
-
-class ShortRope(nn.Module):
-
-    def __init__(self, config):
-        super().__init__()
-        self.n = config.shortrope_length
-        self.dim = config.n_embd
-
-        # Generate freqs of size n rather than full dim
-        self.register_buffer('freq_left', (10000 ** (torch.arange(0, self.n//2).float() / self.n//2)))
-        self.register_buffer('freq_right', (10000 ** (torch.arange(0, self.n//2).float() / self.n//2)))
-
-    def forward(self, x):
-        # Step 1: Get the input tensor shape
-        batch_size, seq_len, _ = x.shape
-
-        # Step 2: Split the input tensor into unrotated and rotated sections
-        x_unrotated = x[..., :-self.n]  # All but the last n dimensions
-        x_rotated = x[..., -self.n:]    # Only the last n dimensions
-
-        # Step 3: Generate rotation frequencies
-        t = torch.arange(self.n, device=x.device)
-        freqs_left = torch.einsum('i,j->ij', t, self.freq_left)
-        freqs_right = torch.einsum('i,j->ij', t, self.freq_right)
-
-        # Calculate how many times to repeat freqs along the sequence length
-        num_repeats = seq_len // self.n + int(seq_len % self.n != 0)
-
-        # Repeat the frequency tensors to match the sequence length
-        freqs_left = freqs_left.repeat(batch_size, num_repeats, 1)
-        freqs_right = freqs_right.repeat(batch_size, num_repeats, 1)
-
-        # Trim the excess elements so the freqs tensors match the sequence length
-        freqs_left = freqs_left[:, :seq_len, :]
-        freqs_right = freqs_right[:, :seq_len, :]
-
-        # Step 4: Process the x_rotated section
-        x_left = x_rotated[..., :self.n//2]
-        x_right = x_rotated[..., self.n//2:]
-
-        # Apply the cosine and sine rotations
-        x_left = x_left * freqs_left.cos() - x_right * freqs_left.sin()
-        x_right = x_left * freqs_right.sin() + x_right * freqs_right.cos()
-
-        # Invert the order of the right tensor's last dimension and negate it
-        x_right = torch.flip(x_right, dims=[-1]) * -1
-
-        # Combine the left and right rotated sections
-        x_rotated = torch.cat([x_left, x_right], dim=-1)
-
-        # Step 5: Combine the rotated and unrotated sections
-        x = torch.cat([x_unrotated, x_rotated], dim=-1)
-
-        return x
+# Variations
+from variations.softmax_variations import Softermax, Constantmax, Constantmax_quan, Strongermax, Polymax, SigSoftmax
+from variations.normalization_variations import LayerNorm, RMSNorm
+from variations.position_encoding_variations import RotaryEmbedding, ShortRope
+from variations.activation_variations import SquaredReLU
 
 
 class CausalSelfAttention(nn.Module):
@@ -403,13 +110,6 @@ class CausalSelfAttention(nn.Module):
         y = self.resid_dropout(self.c_proj(y))
         return y
 
-# Activation variations
-class SquaredReLU(nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    def forward(self, x):
-        return torch.pow(torch.relu(x), 2)
 
 class MLP(nn.Module):
